uploading the profile picture of user
